refactor(cosmos): move rpc and proto into cosmos/ (#442)
