+++ host/cosmos/proto/buf.gen.gogo.yaml
